Fix for compilation issue

+++ snxxx/libese-teq1/nxp-ese/utils/ese_config.h
@@ -68,7 +68,8 @@ class EseConfig {
   static unsigned getUnsigned(const std::string& key);
   static unsigned getUnsigned(const std::string& key, unsigned default_value);
   static unsigned long getUnsignedLong(const std::string& key);
-  static unsigned long getUnsignedLong(const std::string& key, unsigned default_value);
+  static unsigned long getUnsignedLong(const std::string& key,
+                                       unsigned long default_value);
 
  private:
   static EseConfig& getInstance();
